@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "nonebot_plugin_emojimix"
-version = "0.4.0"
+version = "0.4.1"
 description = "Nonebot2 plugin for emojimix"
 authors = ["meetwq <meetwq@gmail.com>"]
 license = "MIT"
@@ -9,22 +9,22 @@ homepage = "https://github.com/noneplugin/nonebot-plugin-emojimix"
 repository = "https://github.com/noneplugin/nonebot-plugin-emojimix"
 
 [tool.poetry.dependencies]
-python = "^3.8"
-nonebot2 = "^2.2.0"
+python = "^3.9"
+nonebot2 = "^2.3.0"
 nonebot-plugin-alconna = ">=0.38.0,<1.0.0"
-httpx = ">=0.20.0,<1.0.0"
+httpx = ">=0.26.0,<1.0.0"
 emoji = "^2.0.0"
 
 [tool.poetry.group.dev.dependencies]
 
 [tool.pyright]
-pythonVersion = "3.8"
+pythonVersion = "3.9"
 pythonPlatform = "All"
 typeCheckingMode = "basic"
 
 [tool.ruff]
 line-length = 88
-target-version = "py38"
+target-version = "py39"
 
 [tool.ruff.lint]
 select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
